+++ Player/Tech/HLS/HLSNativeWarning.swift
@@ -6,16 +6,25 @@
 //  Copyright © 2018 emp. All rights reserved.
 //
 
-import Foundation
+import AVFoundation
+
 
 public enum HLSNativeWarning: WarningMessage {
-    case seekableRangesEmpty(source: MediaSource?)
+    case seekableRangesEmpty
+    case discontinuousSeekableRanges(seekableRanges: [CMTimeRange])
+    case seekTimeBeyondLivePoint(timestamp: Int64, livePoint: Int64)
+    
+    case invalidStartTime(startTime: Int64, seekableRanges: [CMTimeRange])
 }
 
 extension HLSNativeWarning {
     public var message: String {
         switch self {
-        case .seekableRangesEmpty(source: let source): return "Seekable ranges empty for sourceId \(source?.playSessionId)"
+        case .seekableRangesEmpty: return "Seekable ranges was empty"
+        case .discontinuousSeekableRanges(seekableRanges: let ranges): return "Seekable ranges contain discontinuity \(ranges)"
+        case .seekTimeBeyondLivePoint(timestamp: let timestamp, livePoint: let live): return "Requested seek time \(timestamp) was beyond live point \(live)"
+            
+        case .invalidStartTime(startTime: let time, seekableRanges: let ranges): return "Invalid start time, \(time) set beyond seekable ranges, \(ranges)"
         }
     }
 }