+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 #### Features
 * `EMP-10646` `ExposureContext` exposes cached server time
 * Added support for `onWarning` events
+* Exposed `isMuted` and `volumne` apis
 
 #### Changes
 * `EMP-10852` API changes to `MediaPlayback.currentTime`. Property renamed to `MediaPlayback.playheadPosition` which better reflects the actual position of the playback.