+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # CHANGELOG
 
-* `0.76.x` Releases - [0.76.0](#0760)
+* `0.77.x` Releases - [0.77.0](#0770)
 * `0.73.x` Releases - [0.73.0](#0730)
 * `0.72.x` Releases - [0.72.0](#0720)
 * `0.2.x` Releases - [0.2.0](#020)
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015)
 
-## 0.76.0
+## 0.77.0
 
 #### Features
 * `EMP-10646` `ExposureContext` exposes cached server time
@@ -19,6 +19,7 @@
 * Improvements in how `HLSNative` handles `StartTime`.
 * Event callbacks now return `Player` instead of `Tech` directly.
 * Seekable and buffered timeranges now return `[CMTimeRange]`
+* Several improvements to the startup and loading process for `HLSNative`
 
 ## 0.73.0